Add command dispatcher for developer commands
